+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `ClassificationReport` with support for binary, multiclass, and multilabel classification tasks ([#3116](https://github.com/Lightning-AI/torchmetrics/pull/3116))
+
+
 - Added CRPS in regression domain ([#3024](https://github.com/Lightning-AI/torchmetrics/pull/3024))